:sparkles: add etc search filter

+++ app/src/main/java/com/wafflestudio/snutt2/lib/network/dto/PostSearchQueryParams.kt
@@ -16,7 +16,7 @@ data class PostSearchQueryParams(
     @Json(name = "department") val department: List<String>? = null,
     @Json(name = "category") val category: List<String>? = null,
     @Json(name = "etc") val etc: List<String>? = null,
-    @Json(name = "time_mask") val time_mask: List<Long>? = null,
+    @Json(name = "time_mask") val time_mask: List<Int>? = null,
     @Json(name = "offset") val offset: Long? = null,
     @Json(name = "limit") val limit: Long? = null
 )
